refactor repositoryProtocol, remoteDataSourceProtocol and localeDataSourceProtocol
