+++ Shared/Core/Data/Repository/Protocol/UserRepositoryProtocol.swift
@@ -1,0 +1,13 @@
+//
+//  UserRepositoryProtocol.swift
+//  RiMovie (iOS)
+//
+//  Created by Ari Supriatna on 05/01/21.
+//
+
+protocol UserRepositoryProtocol {
+  
+  func getCertificates() -> [CertificateModel]
+  func getMyClass() -> [ClassDicodingModel]
+  
+}
